chart_gen_py/pyproject.toml
[project]
name = "py-euch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.0",
    "python-constraint2>=2.1.0",
]
